Create Model base class
The previous Model and TraitModel classes are now called SpExModel and
TraitModel, which now derive from the Model base class. Much of the
shared code between the model classes have been moved to the Model
base class. There still remains code deplication that needs to be fixed
in the future.

The code compiles and appears to run properly. However, the results do
not match the results from previous versions, given the specific seed
number of 1979. I am looking into why.