Create MCMC base class and make SpExMCMC derive it
MCMC now contains many of the code shared by SpExMCMC and TraitMCMC.
TraitMCMC has not yet been updated, but will be in a future commit.
One important change is that MCMC now has a run() method for running
the chain. It used to be in the class' constructor. The file main.cpp
has been modified to call run() on SpExModel.